v4.3.3 [deploy:worker]

+++ database/migrations/2018_09_14_155616_remove_rn_can_approve_careplans.php
@@ -13,9 +13,12 @@ class RemoveRnCanApproveCareplans extends Migration
      */
     public function up()
     {
-        Schema::table('cpm_settings', function (Blueprint $table) {
-            $table->dropColumn('rn_can_approve_careplans');
-        });
+        if (Schema::hasColumn('cpm_settings', 'rn_can_approve_careplans')) {
+            Schema::table('cpm_settings', function (Blueprint $table) {
+                $table->dropColumn('rn_can_approve_careplans');
+            });
+        }
+
     }
 
     /**
@@ -25,8 +28,10 @@ class RemoveRnCanApproveCareplans extends Migration
      */
     public function down()
     {
-        Schema::table('cpm_settings', function (Blueprint $table) {
-            $table->boolean('rn_can_approve_careplans')->default(0);
-        });
+        if ( ! Schema::hasColumn('cpm_settings', 'rn_can_approve_careplans')) {
+            Schema::table('cpm_settings', function (Blueprint $table) {
+                $table->boolean('rn_can_approve_careplans')->default(0);
+            });
+        }
     }
 }
